next-round: add autosave implementation
There's no way to set the autosave-dir yet, so this dead code is
preparation for a future commit which allows the user to set the
directory. I've decided not to expose do-autosave for now, which means
that we will _not_ create a save when first enabling autosave (contrary
to the issue notes [1]).

[1]: https://github.com/benknoble/frosthaven-manager/issues/121